// x-pack/platform/plugins/shared/alerting/server/application/gaps/types/intervals.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0; you may not use this file except in compliance with the Elastic License
 * 2.0.
 */

export interface Interval {
  gte: Date;
  lte: Date;
}

export interface StringInterval {
  gte: string;
  lte: string;
}